move guard files out of core module folder
